Rename message to portal_messages.
