+++ client/src/components/Workbench/Workbench.test.js
@@ -14,7 +14,7 @@ import {
 } from '../../redux/reducers/tickets.reducers';
 import { initialState as authenticatedUser } from '../../redux/reducers/authenticated_user.reducer';
 import { initialState as systemMonitor } from '../../redux/reducers/systemMonitor.reducers';
-import { initialIntroMessages as introMessages } from '../../redux/reducers/message.reducers';
+import { initialIntroMessages as introMessages } from '../../redux/reducers/portalMessages.reducers';
 import { initialSystemState as systems } from '../../redux/reducers/datafiles.reducers';
 
 import * as introMessageText from '../../constants/messages';
