+++ client/src/redux/reducers/index.js
@@ -16,7 +16,7 @@ import authenticatedUser from './authenticated_user.reducer';
 import { pushKeys } from './systems.reducers';
 import notifications from './notifications.reducers';
 import workbench from './workbench.reducers';
-import { introMessages, customMessages } from './message.reducers';
+import { introMessages, customMessages } from './portalMessages.reducers';
 import { onboarding } from './onboarding.reducers';
 import projects from './projects.reducers';
 import { users } from './users.reducers';
