--- server/portal/apps/message/apps.py
@@ -1,6 +1,0 @@
-from django.apps import AppConfig
-
-
-class MessageConfig(AppConfig):
-    name = 'message'
-    app_label = 'message'
