+++ server/portal/apps/portal_messages/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class PortalMessagesConfig(AppConfig):
+    name = 'portal_message'
+    app_label = 'portal_message'
