+++ server/portal/apps/portal_messages/views.py
@@ -1,5 +1,5 @@
 """
-.. :module: apps.message.views
+.. :module: apps.portal_messages.views
    :synopsis: Views to handle read/unread status of IntroMessages
               and generate CustomMessages
 """
@@ -7,7 +7,7 @@
 import logging
 from portal.views.base import BaseApiView
 from django.http import JsonResponse
-from portal.apps.message.models import IntroMessages, CustomMessages, CustomMessageTemplate
+from portal.apps.portal_messages.models import IntroMessages, CustomMessages, CustomMessageTemplate
 from django.contrib.auth.decorators import login_required
 from django.utils.decorators import method_decorator
 import json
