+++ server/portal/settings/unit_test_settings.py
@@ -85,7 +85,7 @@ INSTALLED_APPS = [
     'portal.apps.googledrive_integration',
     'portal.apps.datafiles',
     'portal.apps.projects',
-    'portal.apps.message',
+    'portal.apps.portal_messages',
 
 ]
